moved pydantic classes to the utils folder
